+++ dashboard/auth/exceptions.py
@@ -1,0 +1,2 @@
+class AuthenticationFailed(Exception):
+    pass
