+++ dashboard/history/service.py
@@ -1,11 +1,35 @@
-# -*- coding: utf-8 -*-
-from collections import namedtuple
+from collections.__init__ import namedtuple
 
-from dashboard.exceptions import PageOutOfRange
-from dashboard.model import ZuulBuildSet
+from dashboard.history.exceptions import PageOutOfRange
+from dashboard.history.model import ZuulBuildSet
 
 Paginator = namedtuple("Paginator", ["pages", "previous_page",
                                      "next_page", "current_page"])
+
+
+def pagination(number_of_buildsets, page, per_page, page_links):
+    if number_of_buildsets < 0 or page < 1 or per_page < 1 or page_links < 0:
+        raise ValueError
+
+    number_of_pages = int((number_of_buildsets + per_page - 1) / per_page)
+
+    if number_of_pages < 1:
+        return Paginator(pages=[], previous_page=None,
+                         next_page=None, current_page=page)
+
+    previous_page = None if page == 1 else page - 1
+    next_page = None if page == number_of_pages else page + 1
+
+    pages = [p for p in range(page - page_links, page + page_links + 1)
+             if 0 < p <= number_of_pages]
+
+    if pages[0] != 1:
+        pages = [1, None] + pages
+
+    if pages[-1] != number_of_pages:
+        pages.extend([None, number_of_pages])
+
+    return Paginator(pages, previous_page, next_page, page)
 
 
 class BuildSetsPaginated:
@@ -41,28 +65,3 @@ class BuildSetsFiltered(BuildSetsPaginated):
 
     def _create_query(self):
         return ZuulBuildSet.get_filtered(self.pipeline, self.branch, self.build)
-
-
-def pagination(number_of_buildsets, page, per_page, page_links):
-    if number_of_buildsets < 0 or page < 1 or per_page < 1 or page_links < 0:
-        raise ValueError
-
-    number_of_pages = int((number_of_buildsets + per_page - 1) / per_page)
-
-    if number_of_pages < 1:
-        return Paginator(pages=[], previous_page=None,
-                         next_page=None, current_page=page)
-
-    previous_page = None if page == 1 else page - 1
-    next_page = None if page == number_of_pages else page + 1
-
-    pages = [p for p in range(page - page_links, page + page_links + 1)
-             if 0 < p <= number_of_pages]
-
-    if pages[0] != 1:
-        pages = [1, None] + pages
-
-    if pages[-1] != number_of_pages:
-        pages.extend([None, number_of_pages])
-
-    return Paginator(pages, previous_page, next_page, page)
