+++ dashboard/history/tests/test_buildsets.py
@@ -2,24 +2,24 @@
 import unittest
 from unittest.mock import MagicMock, patch
 
-from dashboard.exceptions import PageOutOfRange
-from dashboard.history import BuildSetsPaginated
+from dashboard.history.exceptions import PageOutOfRange
+from dashboard.history.service import BuildSetsPaginated
 
 
 class TestBuildSets(unittest.TestCase):
-    @patch('dashboard.model.ZuulBuildSet.get_for_pipeline')
+    @patch('dashboard.history.model.ZuulBuildSet.get_for_pipeline')
     def test_create_buildsets_history_object(self, get_for_pipeline):
         buildsets = BuildSetsPaginated(pipeline="foo", per_page=20)
 
         self.assertEqual(buildsets.per_page, 20)
 
-    @patch('dashboard.model.ZuulBuildSet.get_for_pipeline')
+    @patch('dashboard.history.model.ZuulBuildSet.get_for_pipeline')
     def test_create_query(self, get_for_pipeline):
         BuildSetsPaginated(pipeline="foo", per_page=20)
 
         get_for_pipeline.assert_called_with("foo")
 
-    @patch('dashboard.model.ZuulBuildSet.get_for_pipeline')
+    @patch('dashboard.history.model.ZuulBuildSet.get_for_pipeline')
     def test_fetch_raises_when_page_is_out_of_range(self, get_for_pipeline):
         get_for_pipeline.return_value = [
             'first_element', 'second_element', 'third_element']
@@ -29,7 +29,7 @@ class TestBuildSets(unittest.TestCase):
         with self.assertRaises(PageOutOfRange):
             buildsets.fetch_page(2)
 
-    @patch('dashboard.model.ZuulBuildSet.get_for_pipeline')
+    @patch('dashboard.history.model.ZuulBuildSet.get_for_pipeline')
     def test_fetch_set_correct_page(self, get_for_pipeline):
         data = ['first_element', 'second_element', 'third_element']
