+++ dashboard/status/service.py
@@ -2,7 +2,7 @@ import requests
 from flask import current_app
 
 from dashboard.config import config
-from dashboard.exceptions import RemoteServerError, PipelineNotFound
+from dashboard.status.exceptions import RemoteServerError, PipelineNotFound
 from dashboard.status.model import PipelineStat, Queue
