+++ dashboard/status/tests/static/status_no_pipelines.json
