+++ tests/__init__.py
@@ -5,7 +5,7 @@ from pony.orm import db_session
 
 from dashboard import db
 
-from tests import fixtures
+from dashboard.auth.tests import fixtures
 
 
 class IntegrationTestCase(unittest.TestCase):
@@ -16,8 +16,3 @@ class IntegrationTestCase(unittest.TestCase):
     @classmethod
     def tearDownClass(cls):
         db.db.disconnect()
-
-    @classmethod
-    @db_session
-    def seed_database(cls):
-        fixtures.seed_database()
