--- tests/static/test_users.yml
@@ -1,6 +1,0 @@
----
-admins:
-  - test@email.test
-  - super@user.test
-  - me@example.test
-  - admin@acid.test
