+++ tox.ini
@@ -11,7 +11,7 @@ commands = make install-dev
 [testenv]
 basepython = python3
 setenv =
-  SETTINGS_PATH={toxinidir}/tests/static/test_settings.yml
+  SETTINGS_PATH={toxinidir}/settings_test.yml
   FLASK_ENV=testing
   DEBUG=0
   TESTING=1
