+++ users.yml
@@ -1,3 +1,6 @@
 ---
 admins:
+  - test@email.test
+  - super@user.test
+  - me@example.test
   - admin@acid.test
